requirements.txt
streamlit
fastapi
uvicorn
opencv-python-headless
numpy
pillow
requests
python-multipart

